third_party/WebKit/LayoutTests/external/wpt/xhr/event-timeout-order-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: event - timeout (order of events) assert_equals: expected "upload.loadstart(0,12,true)" but got "upload.loadstart(0,0,false)"
Harness: the test ran to completion.
